requirements.txt
paho-mqtt~=1.6.1
flask~=2.3.2
whereami
gmaps
IPython
ipywidgets
folium~=0.14.0
geopy~=2.3.0
python-telegram-bot
nanodb
folium